LPS-179567 Extract CPInstancePriceListsScreenNavigationCategory methods to a CPInstancePriceListsScreenNavigationEntry component and extends CPInstancePriceListsScreenNavigationCategory in it
