+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/taglib/servlet/taglib/CPInstancePriceListsScreenNavigationCategory.java
@@ -14,31 +14,13 @@
 
 package com.liferay.commerce.pricing.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.price.list.portlet.action.CommercePriceListActionHelper;
-import com.liferay.commerce.price.list.service.CommercePriceEntryService;
-import com.liferay.commerce.pricing.web.internal.display.context.CPInstanceCommercePriceEntryDisplayContext;
-import com.liferay.commerce.product.model.CPInstance;
-import com.liferay.commerce.product.portlet.action.ActionHelper;
 import com.liferay.commerce.product.servlet.taglib.ui.constants.CPInstanceScreenNavigationConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
-import com.liferay.item.selector.ItemSelector;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -47,22 +29,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=30",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=30",
+	service = ScreenNavigationCategory.class
 )
 public class CPInstancePriceListsScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<CPInstance> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return "price-lists";
-	}
-
-	@Override
-	public String getEntryKey() {
 		return "price-lists";
 	}
 
@@ -71,7 +45,7 @@ public class CPInstancePriceListsScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, "price-lists");
+		return language.get(resourceBundle, "price-lists");
 	}
 
 	@Override
@@ -80,67 +54,7 @@ public class CPInstancePriceListsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_CP_INSTANCE_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, CPInstance cpInstance) {
-		if (cpInstance == null) {
-			return false;
-		}
-
-		return true;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		try {
-			CPInstanceCommercePriceEntryDisplayContext
-				cpInstanceCommercePriceEntryDisplayContext =
-					new CPInstanceCommercePriceEntryDisplayContext(
-						_actionHelper, _commercePriceEntryService,
-						_commercePriceListActionHelper, httpServletRequest,
-						_itemSelector);
-
-			httpServletRequest.setAttribute(
-				WebKeys.PORTLET_DISPLAY_CONTEXT,
-				cpInstanceCommercePriceEntryDisplayContext);
-		}
-		catch (Exception exception) {
-			_log.error(exception);
-		}
-
-		_jspRenderer.renderJSP(
-			_setServletContext, httpServletRequest, httpServletResponse,
-			"/commerce_price_lists/cp_instance" +
-				"/cp_instance_commerce_price_lists.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		CPInstancePriceListsScreenNavigationCategory.class);
-
 	@Reference
-	private ActionHelper _actionHelper;
-
-	@Reference
-	private CommercePriceEntryService _commercePriceEntryService;
-
-	@Reference
-	private CommercePriceListActionHelper _commercePriceListActionHelper;
-
-	@Reference
-	private ItemSelector _itemSelector;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.pricing.web)"
-	)
-	private ServletContext _setServletContext;
+	protected Language language;
 
 }
